Fix for GH-11
I did try making them private via double underscore, but that broke unit testing. Adding a way for unit tests to set state defeats the purpose of private IMHO.

I did prefix them with underscores, and ensure they aren't referenced outside their package or  the test code. This did require adding a getter (that returns a deepcopy) and setter for user_variables (which also duplicated validation on the setter) and adding a getter for memval that returns the variable so it can be bound to its Label.